fix(studio): ingest the live 2D-sources options from the renderer snapshot
The Side/Back buttons (and the spatialize toggle, object-generator selector,
phantom controls, output channel mapping, virtual bed) never reflected the
engine's actual state: the values were dropped twice on the way to the UI.

1. The Tauri mirror (RendererDomainState / AppState) never carried
   channelRenderMode, surroundPlacement, objectGenerator*, phantom*,
   outputChannelMapping* or virtualBed — the snapshot keys were silently
   discarded at the re-model. They now ride a single LiveOptionsState struct
   flattened into both, so future options are one field away.

2. The only JS ingestion of those fields lived in runtime-audio-state.js,
   which has been orphaned since 4ba0330 (April) inlined a partial copy of it
   into init.js — every ingestion added to that module afterwards was dead
   code. The missing fields (plus latencySmoothedMs) are now ingested in
   applyInitState itself, the virtual-bed materialise one-shot included, and
   the dead module is deleted so state ingestion has exactly one home.

Visible symptom fixed: launching a film against an engine whose in-memory
placement differed from the UI default showed 'Side' active while the bed
rendered at the back, until a manual toggle round-trip.
