+++ omniphony-studio/src-tauri/src/osc_listener.rs
@@ -7,8 +7,8 @@ use tokio::sync::mpsc::UnboundedReceiver;
 
 use crate::app_state::OutputDeviceOption;
 use crate::app_state::{
-    AppState, DistanceDiffuse, Meter, RenderBackendState, RoomRatio, SpreadState, VbapCartesian,
-    VbapPolar,
+    AppState, DistanceDiffuse, LiveOptionsState, Meter, RenderBackendState, RoomRatio, SpreadState,
+    VbapCartesian, VbapPolar,
 };
 use crate::layouts::{Layout, Speaker};
 use crate::osc_parser::{
@@ -142,6 +142,10 @@ struct RendererDomainState {
     vbap_cartesian: Option<VbapCartesian>,
     vbap_polar: Option<VbapPolar>,
     render_backend_state: Option<RenderBackendState>,
+    /// The live 2D-sources / routing options, collected by key (flatten) and
+    /// mirrored into `AppState` verbatim — see [`LiveOptionsState`].
+    #[serde(flatten)]
+    live_options: LiveOptionsState,
 }
 
 #[derive(serde::Deserialize)]
@@ -641,6 +645,10 @@ fn apply_renderer_domain_state(s: &mut AppState, value: &str) -> bool {
     if let Some(render_backend_state) = parsed.render_backend_state {
         s.render_backend_state = render_backend_state;
     }
+    // The renderer domain always carries the full option set, so mirror it
+    // wholesale (an explicit `virtualBed: null` must reach the UI — it means
+    // "no saved bed", which triggers the one-shot canonical-bed materialise).
+    s.live_options = parsed.live_options;
     true
 }
